runtime: share init data setup code
Move init data setup code such that it can be used by other hypervisors.
